refactor: break up LandingPage into modular components
