+++ client/src/pages/Landing/Footer/Footer.tsx
@@ -1,0 +1,9 @@
+const Landing = () => {
+  return (
+    <footer className="text-center py-6 border-t border-light-cont-s dark:border-dark-cont-s text-light-text-sec dark:text-dark-text-sec transition-colors">
+      © {new Date().getFullYear()} Lambda Lens. All rights reserved.
+    </footer>
+  );
+};
+
+export default Landing;
